Problem: store is using one thread for both writing and reading
which is very slow and make consumer wait for writing to complete.
Also we are not utilizing leveldb primitives like batch writing and iterating

Solution: separate store into write and read actors which use the same underlying leveldb.

Store is using batch for writing (up to 1000) and only ack the last message per subject/address of a batch