Problem: It hard to write deterministic tests
One issue are HEAD messages which are send at certain interval. Also if
there're both store and producer nodes either of them might send a
answer.

Solution: Create a test peer that listen to message which can be obtains
for assertions. Also it brings convince methods to send messages and
assert responses.